javalinapp: move jsonmapper to externalobject
